Refactor: move input types to separate module
