+++ packages/html/src/lib.rs
@@ -17,9 +17,11 @@ mod elements;
 mod events;
 pub mod geometry;
 mod global_attributes;
+mod input;
 #[cfg(feature = "wasm-bind")]
 mod web_sys_bind;
 
 pub use elements::*;
 pub use events::*;
 pub use global_attributes::*;
+pub use input::*;
